% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_data_hydrobio.R
\name{telecharger_listes}
\alias{telecharger_listes}
\title{Title}
\usage{
telecharger_listes(
  code_departement,
  code_eqb = c(Poissons = 4, Diatomées = 10, Macroinvertébrés = 13, Macrophytes = 27)
)
}
\arguments{
\item{code_eqb}{}
}
\description{
Title
}
